requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
twilio==6.51.1
urllib3==1.26.3
